Fix stack bar flicker when hovering, fixes #51 (#53)
By moving the `on('mouseout')` handler to the stacks-wrapper element, we
avoid firing it when moving between frames. This way, the stack pointer
animates more smoothly when moving horizontally from one frame to the
next, hovering in between two frames doesn't keep reverting to the
`selectedNode`.

The 'mouseout' event bubbles for some reason so we have to check that it
fired on the stacks-wrapper element and not a child element. Gotta love
the DOM :fireworks:

This also moves the `on('mouseover')` handler to the stacks-wrapper element,
which saves some time by not reattaching event listeners on every render.