Store search state in URL hash. (#27)
When searching, the query is added to the hash as `&search=query here`. If the search query changes from something like `app` to `app.get`, it's assumed that the previous search was triggered while the user was still typing, and the hash is updated without creating a new history entry. This way you don't have to flick through a bunch of partial queries to go back in the history if you type slow.

This emits `search` and `clearSearch` events from UI so the SearchBox input can update, and moves the FlameGraph search() calls to also use those events.