Avoid rerendering flame graph if size hasn't changed (#36)
`cellHeight()` and `width()` immediately rerender the graph, so doing it
unconditionally is quite expensive.

With this patch we rerender far less frequently, and an update just
causes 1 render instead of 3(!). If the size changed, we still do 2
renders, because of how d3-fg works … maybe we could use
`requestAnimationFrame` in d3-fg to batch updates, but that would
likely be a breaking change, and we rely on filter updates etc. being
processed synchronously in Flame.