Add --dest flag. (#31)
This adds a `dest: 'path str'` option to Flame so the output location of the graphs and log data is configurable. With `dest: 'test-output-destination'`, files are placed in `test-output-destination/{pid}.clinic-flame`. This needed a change in `getLoggingPaths()`, that was already applied in clinic-common, so I figured this would be a good time to switch to clinic-common's implementation.

This requires https://github.com/nearform/node-clinic-common/pull/3 so clinic-common can support our Flame-specific files, tests pass with that patch.